--- src/main/java/com/bookmate/bookmate/book/entity/Book.java
@@ -1,66 +1,0 @@
-package com.bookmate.bookmate.book.entity;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@Entity
-@EntityListeners(AuditingEntityListener.class)
-@Table(name = "book")
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class Book {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(nullable = false, unique = true, length = 13)
-  private String isbn13;
-
-  @Column(nullable = false)
-  private String title;
-
-  @Column(nullable = false)
-  private String author;
-
-  @Column(nullable = false)
-  private String publisher;
-
-  @Column(nullable = false, name = "cover_url")
-  private String coverUrl;
-
-  @Column(nullable = false, name = "published_date")
-  private LocalDate publishedDate;
-
-  @Column(nullable = false)
-  private String description;
-
-  @OneToMany(mappedBy = "book", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-  private List<SaveBook> saveBooks;
-
-  @CreatedDate
-  @Column(updatable = false)
-  private LocalDateTime createdAt;
-
-  @LastModifiedDate
-  private LocalDateTime updatedAt;
-}
