--- src/main/java/com/bookmate/bookmate/book/repository/BookRepository.java
@@ -1,10 +1,0 @@
-package com.bookmate.bookmate.book.repository;
-
-import com.bookmate.bookmate.book.entity.Book;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BookRepository extends JpaRepository<Book, Long> {
-
-}
