+++ src/main/java/com/bookmate/bookmate/main/dto/AladinApiResponseDto.java
@@ -1,4 +1,4 @@
-package com.bookmate.bookmate.book.dto;
+package com.bookmate.bookmate.main.dto;
 
 
 import lombok.AllArgsConstructor;
