+++ src/main/java/com/bookmate/bookmate/main/service/AladinApiService.java
@@ -1,7 +1,7 @@
-package com.bookmate.bookmate.book.service;
+package com.bookmate.bookmate.main.service;
 
 
-import com.bookmate.bookmate.book.dto.AladinApiResponseDto;
+import com.bookmate.bookmate.main.dto.AladinApiResponseDto;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.ArrayList;
